breaking: fix: kcp2k V1.16 - fix: SendUnreliable respects ArraySegment.Offset - fix: potential bug with negative length (see PR #2) - breaking: removed pause handling because it's not necessary for Mirror anymore

+++ Assets/Mirror/Runtime/Transport/KCP/kcp2k/kcp/Kcp.cs
@@ -27,7 +27,7 @@ namespace kcp2k
         public const int INTERVAL = 100;
         public const int OVERHEAD = 24;
         public const int FRG_MAX = byte.MaxValue;  // kcp encodes 'frg' as byte. so we can only ever send up to 255 fragments.
-        public const int DEADLINK = 20;            // default maximum amount of 'xmit' retransmissions until a message is considered lost
+        public const int DEADLINK = 20;            // default maximum amount of 'xmit' retransmissions until a segment is considered lost
         public const int THRESH_INIT = 2;
         public const int THRESH_MIN = 2;
         public const int PROBE_INIT = 7000;        // 7 secs to probe window size
@@ -65,7 +65,7 @@ namespace kcp2k
         internal bool updated;
         internal uint ts_probe;      // timestamp probe
         internal uint probe_wait;
-        internal uint dead_link;     // maximum amount of 'xmit' retransmissions until a message is considered lost
+        internal uint dead_link;     // maximum amount of 'xmit' retransmissions until a segment is considered lost
         internal uint incr;
         internal uint current;       // current time (milliseconds). set by Update.
 
@@ -535,7 +535,8 @@ namespace kcp2k
                 size -= OVERHEAD;
 
                 // enough remaining to read 'len' bytes of the actual payload?
-                if (size < len || len < 0) return -2;
+                // note: original kcp casts uint len to int for <0 check.
+                if (size < len || (int)len < 0) return -2;
 
                 if (cmd != CMD_PUSH && cmd != CMD_ACK &&
                     cmd != CMD_WASK && cmd != CMD_WINS)
